break out some common functions and interfaces
