+++ src/utils/CaminoExporter.ts
@@ -2,14 +2,7 @@ import { Pathway } from 'pathways-model';
 import { Criteria } from 'criteria-model';
 import { v4 as uuidv4 } from 'uuid';
 import { deepCopyPathway } from './nodeUtils';
-
-// interface purely for intermediate working objects
-interface IncludedCqlLibraries {
-  [id: string]: {
-    cql: string;
-    version: string;
-  };
-}
+import { constructCqlLibrary, IncludedCqlLibraries } from './export';
 
 export class CaminoExporter {
   pathway: Pathway;
@@ -84,40 +77,4 @@ export class CaminoExporter {
 
     return this.pathway;
   }
-}
-
-function constructCqlLibrary(
-  libraryName: string,
-  includedCqlLibraries: IncludedCqlLibraries,
-  referencedDefines: Record<string, string>,
-  builderDefines: Record<string, string>
-): string {
-  const includes = Object.entries(includedCqlLibraries)
-    .map(([name, details]) => `include "${name}" version '${details.version}' called ${name}\n\n`)
-    .join('');
-  const definesList = Object.entries(referencedDefines).map(
-    ([name, srcLibrary]) => `define "${name}": ${srcLibrary}.${name}\n\n`
-  );
-
-  Object.entries(builderDefines).forEach(([statement, cql]) =>
-    definesList.push(`define "${statement}": ${cql}\n\n`)
-  );
-
-  const defines = definesList.join('');
-
-  // NOTE: this library should use the same FHIR version as all referenced libraries
-  // and if we want to run it in cqf-ruler, as of today that needs to be FHIR 4.0.1 (NOT 4.0.0)
-  const libraryCql = `
-library ${libraryName} version '1.0'
-
-using FHIR version '4.0.1'
-
-${includes}
-
-context Patient
-
-${defines}
-`;
-
-  return libraryCql;
 }
